Changed blueprint probability to decimal so it matches 038b155960e5ad554fb81f1336e6e21b1c9f2c30 This is purely done for ease of comparison, or the xml would differ for all probabilities on double/decimal

+++ tools/XmlGenerator/Datafiles/Blueprints.cs
@@ -450,7 +450,7 @@ namespace EVEMon.XmlGenerator.Datafiles
                                       (requirement.ActivityID == (int)BlueprintActivity.Invention ||
                                        requirement.ActivityID == (int)BlueprintActivity.ReverseEngineering)))
             {
-                blueprint.InventionTypeIDs.Add(requirement.RequiredTypeID, requirement.Probability.GetValueOrDefault());
+                blueprint.InventionTypeIDs.Add(requirement.RequiredTypeID, (decimal)requirement.Probability.GetValueOrDefault());
             }
         }
